Remove unnecessary code, add ':echo()' for toy 'eval()', remove ':eval()' parsing from this package
The idea is to make the GUI like a REPL. Calling ':eval("some_func")'
lets me call "some_func" defined in the application.

Entire 'get_arg()' function and the part of 'evaluate()' dealing with
':eval()' was a waste of time to write. Remove all of this to (and its
tests) to 'unnecessary_code.py'.

'eval()' is pointless in my pygameapi package because ':eval('__name__')
is pygameapi.pygameapi, not __main__. Therefore, I cannot call
application functions from the GUI command line, and that defeats the
point. I included a snipped in the docstring showing how to implement
':eval' in the application.

I made a simple ':echo' command. Call Python's built-in 'eval()', and
let this return whatever 'eval()' returned. Handle exceptions in a
try/except by returning the exception error message. No reason for the
application to halt on an 'echo()' command line typo!

Since I pass the argument of 'echo' to 'eval', the command line syntax
is pleasantly forgiving and intuitive (echo 1+1 outputs 2, echo(1+1)
outputs 2, and echo('1+1') outputs '1+1'.

Python user's will quickly realize from experimenting in the GUI that
':echo(expression)' is really 'eval(expression)', with 'eval' called in
the context of this package, not the application.